Add files via upload
Updated the audio visualizer look and feel. Error pertaining to the recreating of an HTMLMediaElement resolved. Added new menu tag containing the new location of where playlists are held and the upload button. Added a presets feature that I will continue to expand on in the future. Song title, artist, and album cover display added. More features will be added in the future.